Fixed wrong grid dimension allocation (performing integer division and using ceiling on that -> performing float division)

+++ Lab5/src/src_cuda/nbody.cu
@@ -74,7 +74,7 @@ int main(const int argc, const char** argv) {
 
 	// Set geometry
 	dim3 block(THREADS_PER_BLOCK, 1, 1);
-	dim3 grid((int)(ceil(nBodies/THREADS_PER_BLOCK)), 1, 1);
+	dim3 grid((int)(ceil((float)nBodies/THREADS_PER_BLOCK)), 1, 1);
 
 	/****************************** Data transfers ******************************/
 	cudaMalloc((void **) &d_buf, bytes);
